+++ src/modules/products/products.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
+import { MySQLPrismaService } from 'src/database/mysql.service';
 import { SQLServerPrismaService } from 'src/database/sqlserver.service';
 
 @Injectable()
 export class ProductsService {
-  constructor(private readonly sql: SQLServerPrismaService) {}
-
+constructor(private readonly sql: SQLServerPrismaService, private readonly mysql: MySQLPrismaService) { }
   async getProducts(page = 1, limit = 10) {
     const skip = (page - 1) * limit;
 
@@ -27,6 +27,30 @@ export class ProductsService {
 
     return {
       data: products,
+      total,
+      page,
+      lastPage: Math.ceil(total / limit),
+    };
+  }
+  async getCategorys(page = 1, limit = 15) {
+    const skip = (page - 1) * limit;
+
+    const [categorys, total] = await Promise.all([
+      this.mysql.clcategoria.findMany({
+        skip,
+        take: limit,
+
+        select:{
+          codcat: true,
+          catdes: true,
+
+        }
+      }),
+      this.mysql.clcategoria.count(),
+    ]);
+
+    return {
+      data: categorys,
       total,
       page,
       lastPage: Math.ceil(total / limit),